requirements.txt
streamlit==1.41.1
pandas==2.2.3
pytest==8.3.4
pydantic==2.10.4
pydantic_core==2.27.2
mysql-connector-python==9.1.0
uvicorn==0.34.0
fastapi==0.115.6
fastapi-cli==0.0.7
requests==2.32.3


